requirements.txt
python-mpd2
Adafruit-GPIO
